improved logging: add logback-spring.xml and made HttpLogInterceptor.java safer
