+++ .gitignore
@@ -34,4 +34,6 @@ build/
 /.mvn/wrapper/maven-wrapper.properties
 /mvnw
 /mvnw.cmd
+
 /.env
+/logs/
