+++ src/main/java/com/evgeniyfedorchenko/gptbot/configuration/OkHttpClientConfiguration.java
@@ -1,6 +1,5 @@
 package com.evgeniyfedorchenko.gptbot.configuration;
 
-import com.evgeniyfedorchenko.gptbot.configuration.interceptor.HttpLogInterceptor;
 import com.evgeniyfedorchenko.gptbot.configuration.properties.OkhttpProperties;
 import lombok.AllArgsConstructor;
 import okhttp3.ConnectionPool;
