+++ src/main/resources/application.properties
@@ -51,4 +51,4 @@ spring.data.redis.host=${REDIS_HOST}
 spring.data.redis.password=${REDIS_PASSWORD}
 
 #   Logging
-logging.level.com.evgeniyfedorchenko.gptbot=trace
+logging.file.path=logs/
